requirement.txt
streamlit
torch
transformer